+++ src/defaultEnv.js
@@ -1,7 +1,7 @@
 import E from '@rdfjs/environment'
-import DataFactory from '@rdfjs/environment/DataFactory.js'
-import DatasetFactory from '@rdfjs/environment/DatasetFactory.js'
-import NamespaceFactory from '@rdfjs/environment/NamespaceFactory.js'
+import DataFactory from '@rdfjs/data-model/Factory.js'
+import DatasetFactory from '@rdfjs/dataset/Factory.js'
+import NamespaceFactory from '@rdfjs/namespace/Factory.js'
 
 /**
  * @typedef {import('@rdfjs/environment/Environment.js').Environment<